set destroy action
controller .rbを編集しました。
デストロイ　アクションを追加

+++ app/controllers/prototypes_controller.rb
@@ -24,8 +24,9 @@ class PrototypesController < ApplicationController
 
   def destroy
     prototype = Prototype.find(params[:id])
-    prototype.destroy
+    prototype.destroy if current_user.id == prototype.user_id
   end
+
 
   private
